Crear cuenta empleado

+++ src/main/java/com/reto/usuario/infrastructure/drivenadapter/repository/IRolRepositoryMysql.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRolRepositoryMysql extends JpaRepository<RolEntity, Integer> {
+public interface IRolRepositoryMysql extends JpaRepository<RolEntity, Long> {
 
     RolEntity findByName(String name);
 }
